feat(116): remove circular deps
