+++ package.json
@@ -326,6 +326,7 @@
     "glob": "^7.1.6",
     "husky": "^4.2.3",
     "lint-staged": "^10.0.8",
+    "madge": "^6.1.0",
     "mocha": "^7.1.2",
     "mock-fs": "^4.13.0",
     "parcel": "^1.12.4",
