+++ scripts/find-circular.js
@@ -1,0 +1,5 @@
+const madge = require('madge');
+
+madge('./src/extension.ts').then((res) => {
+  console.log(res.circular());
+})
