+++ src/templates/helpers/isFile.ts
@@ -1,5 +1,4 @@
-import { FileTree } from '../../webviews/Workspace/helpers/getFileTree'
-import { File } from '../../webviews/Workspace/WorkspaceViewProvider.interface'
+import { File, FileTree } from '../../webviews/Workspace/WorkspaceViewProvider.interface'
 
 export const isFile = (item: File | FileTree): item is File => {
   return (item as File).file !== undefined
