+++ src/webviews/Workspace/helpers/compactTree.ts
@@ -1,7 +1,22 @@
 import * as path from 'path'
-import { FileTree, FileTrees } from './getFileTree'
+import { FileTree, FileTrees } from '../WorkspaceViewProvider.interface'
 import { isTreeFolderCompacted } from './isTreeFolderCompacted'
 
+/**
+ * This will compact folders in a similar way to the "Explorer: Compact Folders" setting.
+ * Single child folders will be compressed in a combined tree element. Useful for Java package structures, for example.
+ *
+ * i.e.:
+ *
+ * -Test
+ *  - Subfolder
+ *    - File 1
+ *
+ * Wuld be displayed as
+ *
+ * -Test / Subfolder
+ *  - File 1
+ */
 export const compactTree = (tree: FileTree): FileTree => {
   if (tree.sub.length > 0) {
     tree.sub = tree.sub.reduce((newSubs: FileTrees, curSub: FileTree) => {
