+++ src/webviews/Workspace/helpers/isTreeFolderCompacted.ts
@@ -1,5 +1,5 @@
-import { FileTree } from './getFileTree'
 import * as path from 'path'
+import { FileTree } from '../WorkspaceViewProvider.interface'
 
 export const isTreeFolderCompacted = (tree: FileTree): boolean => {
   return tree.label.includes(path.sep)
